test(dbpull): split tests in multiple files
